pyproject.toml
[project]
name = "cmccm"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "djangorestframework>=3.16.1",
    "django-filter>=25.2",
    "django-cors-headers>=4.9.0",
    "django-allauth[headless,mfa]>=65.14.3",
    "django-anymail[resend]>=14.0",
    "django-storages>=1.14.6",
    "pillow>=12.1.1",
    "psycopg[binary]>=3.3.3",
    "dj-database-url>=3.1.2",
    "python-dotenv>=1.2.1",
    "firebase-admin",
    "supabase>=2.28.2"
]